ui-logicを一つのファイルにまとめた
