--- src/js/game/td-scenes/battle/ui-logic/pilot-button.js
@@ -1,14 +1,0 @@
-// @flow
-
-import type {Command} from "gbraver-burst-core";
-
-/**
- * パイロットボタンが押せるか否かを判定する
- *
- * @param commands プレイヤーが選択可能なコマンド一覧
- * @return 判定結果、trueでボタンが押せる
- */
-export function canPilotButtonPush(commands: Command[]): boolean {
-  const burstCommands = commands.filter(v => v.type === 'PILOT_SKILL_COMMAND');
-  return 1 <= burstCommands.length;
-}
